pyproject.toml
[project]
name = "rag-chatbot-poc"
version = "0.1.0"
description = "This project is a proof of concept for a chatbot application using FastAPI and MongoDB. It demonstrates how to build a simple chatbot that can respond to user queries."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg",
    "bcrypt",
    "cryptography",
    "fastapi",
    "fastapi-sessions",
    "fastapi-sse",
    "greenlet",
    "jinja2",
    "langchain",
    "langchain_openai",
    "motor",
    "pre-commit",
    "pydantic",
    "python-jose",
    "pydantic[email]",
    "pydantic-settings",
    "pyjwt",
    "python-dotenv",
    "python-multipart",
    "passlib",
    "redis",
    "ruff",
    "uvicorn",
]

[dependency-groups]
lint = [
    "ruff>=0.9.10",
]

test = [
    "httpx==0.27.0",
    "polyfactory>=2.19.0",
    "pytest==8.0.2",
    "pytest-asyncio==0.23.5",
]

[tool.pytest.ini_options]
pythonpath = [
    "backend",
]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"
preview = true

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
select = ["E", "F", "DOC"]
ignore = [
  "TRY", "UP006", "BLE", "PGH", "ERA", "E50", "FA1", "INP", "CPY", "D",
  "A001", "A002", "B008", "B904", "ANN201", "ANN001", "ANN003", "ANN002",
  "ANN205", "DOC201","DOC501", "PLR2004", "COM812", "ISC001", "RUF001", "S311",
  "PLR0917", "PLR0913", "G004", "RUF009", "ANN401", "PLR6301", "RET504",
  "PD901", "FBT001", "EM101", "EM102", "PD002", "PLR0914", "RUF002",
  "C901", "PTH118", "PTH120", "PLR0912", "FBT002", "PLR0915", "PLR0911"
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"**/tests/*.py" = ["S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
